requirements/requirements_timeSformer.txt
torch==1.8.1
torchvision==0.9.1
torchvision               0.2.2   
simplejson
einops
timm

psutil
scikit-learn
opencv-python  pip install opencv-python-headless
tensorboard

pyyaml
munch
matplotlib

Step 1:
conda create -n timesformer python=3.7 -y
source activate timesformer

Step 2:
conda install pytorch=1.8.1 torchvision=0.9.1 cudatoolkit=10.2 -c pytorch   THIS WORKS!!!

Step 3:
pip install 'git+https://github.com/facebookresearch/fvcore'

Step 4:
pip install simplejson
pip install einops
pip install timm

Step 5:
conda install av -c conda-forge

Step 6:
pip install psutil
pip install scikit-learn
pip install opencv-python-headless
pip install tensorboard

Step 7:
pip install pyyaml
pip install munch
pip install matplotlib

Step 8:
select correct interpreter (the conda timesformer environment) in vscode

Step 9:
download TimeSformer_divST_8x32_224_k600.pyth


# loader.py: -> put changes into train net
# train_net:
# test_net:
# vit.py: new approach
# meters.py: ##### change from (1,5) def finalize_metrics(self, ks=(1, 2)):
# meters.py: self.video_labels[vid_id] = labels #####[ind]
# checkpoint.py: "cfg": cfg.dump(), ##### 